requirements.txt
botocore
python-dotenv
textract-trp
tabulate
streamlit
boto3
pandas
amazon-textract-response-parser
amazon-textract-caller
langchain_community
langchain
langchain-experimental
langchain-openai
tiktoken
langchain-chroma
amazon-textract-textractor
numpy
unstructured[pdf]
openinference-instrumentation-langchain
arize-phoenix-otel
openpyxl